Add getdependencies.py script for non-CM device/kernel repo additions
This script is intended to be used for adding device/kernel trees and
dependencies that are not hosted on CM github to the local_manifest.
It uses the same file format as getvendorprops.py for defining the repos
to be added. Both the main manifest and the local_manifest are checked
prior to making additions. Configuration files are pulled from the
vendor/pa/dependencies folder. Naming convention for configuration files
is "boardname.dependencies". This file should be called from the product
makefile prior to the "Inherit CM device configuration". Calling syntax
is "GET_DEPENDENCIES := $(shell vendor/pa/tools/getdependencies.py
$(TARGET_PRODUCT)".

Change-Id: I60b034ba8ca387f85985d1ca42d273f66db6288b