Python scripts to modify local_manifest.xml to add/remove projects
addprojects.py
This script uses a file (sys.argv[1]) to add additional projects to the
local_manifest. If the argument is the product name (i.e. pa_maguro), the
prefix will be stripped and the .adds suffix applied (maguro.adds). This
allows $(PRODUCT_NAME) to be used from product makefiles. If no prefix is
found, the file named supplied will be used directly.

removeprojects.py
This script uses a file (sys.argv[1]) to add remove-project elements to
the local_manifest. This allows the removal of projects specified in the
manifest.xml file. This allows addition of replacement project repos to
be added without conflict. The remove-project elements are inserted at the
top of the local_manifest to address potential issues with the replacement
projects being entered before the projects are removed.

Change-Id: I33b1dc05b380334fb1e600d2cf439303aa1770f0